+++ server/form/user_form.go
@@ -1,0 +1,7 @@
+package form
+
+type UserForm struct {
+	Name     string `json:"name" gorm:"type:varchar(255)"`
+	Email    string `json:"email" gorm:"type:varchar(255)"`
+	Password string `json:"password" gorm:"type:varchar(255)"`
+}
